requirements.txt
torch==2.5.1
transformers[timm]==4.46.3
gradio==5.8.0
fastapi==0.115.6
pillow==11.0.0
pydantic==2.10.3
uvicorn==0.32.1
matplotlib==3.9.3
